refactor: Move team-related data to its own package
