--- main/src/io/github/projectpidove/showdown/Gender.scala
@@ -1,5 +1,0 @@
-package io.github.projectpidove.showdown
-
-enum Gender:
-  case Male
-  case Female
