+++ main/src/io/github/projectpidove/showdown/team/Gender.scala
@@ -1,0 +1,5 @@
+package io.github.projectpidove.showdown.team
+
+enum Gender:
+  case Male
+  case Female
