+++ main/src/io/github/projectpidove/showdown/team/Type.scala
@@ -1,4 +1,4 @@
-package io.github.projectpidove.showdown
+package io.github.projectpidove.showdown.team
 
 enum Type:
   case Bug
